+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+- Using michaelhall/http-client for API call to validator service.
 
 ## [2.0.0] - 2018-05-14
 - Updated PHP requirement to 7.1.